requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
geomet==0.2.1.post1
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pydantic==2.11.3
pydantic-core==2.33.1
python-dotenv==1.1.0
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
tenacity==9.1.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0
